readline: use the default emacs keybindings
Now that I actually know emacs keybindings, I think I prefer them for
line-editing in readline. Also, this is the default on pretty much all
servers, and it's annoying to switch between local vi keybindings and
remote emacs keybindings. Better to use the default emacs keybindings
everywhere.

I leave this commented out, and set show-mode-in-prompt on, while I get
used to it. Eventually I will delete both of these options and return to
the defaults.