emacs: explicitly set indent-tabs-mode t
Sigh... better-defaults sets indent-tabs-mode nil, but I depend on using
C styles with tabs for Linux.

I strongly disagree with setting it nil, since that disables any
insertion of tabs for any modes, which breaks how c-default-style is
supposed to work. But apparently tabs are just so evil that the users
deserve to be confused.