emacs: inline better-defaults
I've startd wanting to revert most of the settings in better-defaults,
so I've just inlined the ones I want to keep and I'll stop using the
package.  It didn't really live up to its promise of being
uncontroversial...

Most significant here is that I've stopped using ido-mode.  The
completion system in Emacs 29 is perfectly good on its own!